Restructure code to avoid circular import
